+++ test/clustering/record-linkage/abstract.js
@@ -1,0 +1,22 @@
+/* eslint no-new: 0 */
+/**
+ * Talisman clustering/record-linkage/abstract tests
+ * ==================================================
+ *
+ */
+import assert from 'assert';
+import RecordLinkageClusterer from '../../../src/clustering/record-linkage/abstract';
+
+describe('abstract', function() {
+
+  it('should throw on invalid arguments.', function() {
+
+    assert.throws(function() {
+      new RecordLinkageClusterer(null);
+    }, /params/);
+
+    assert.throws(function() {
+      new RecordLinkageClusterer({}, null);
+    }, /items/);
+  });
+});